Update encode_images.py
Check for encoded files and avoid if script reruns.

+++ encode_images.py
@@ -56,23 +56,24 @@ def main():
     # Optimize (only) dlatents by minimizing perceptual loss between reference and generated images in feature space
     for images_batch in tqdm(split_to_batches(ref_images, args.batch_size), total=len(ref_images)//args.batch_size):
         names = [os.path.splitext(os.path.basename(x))[0] for x in images_batch]
-
-        perceptual_model.set_reference_images(images_batch)
-        op = perceptual_model.optimize(generator.dlatent_variable, iterations=args.iterations, learning_rate=args.lr)
-        pbar = tqdm(op, leave=False, total=args.iterations)
-        for loss in pbar:
-            pbar.set_description(' '.join(names)+' Loss: %.2f' % loss)
-        print(' '.join(names), ' loss:', loss)
-
-        # Generate images from found dlatents and save them
-        generated_images = generator.generate_images()
-        generated_dlatents = generator.get_dlatents()
-        for img_array, dlatent, img_name in zip(generated_images, generated_dlatents, names):
-            img = PIL.Image.fromarray(img_array, 'RGB')
-            img.save(os.path.join(args.generated_images_dir, f'{img_name}.png'), 'PNG')
-            np.save(os.path.join(args.dlatent_dir, f'{img_name}.npy'), dlatent)
-
-        generator.reset_dlatents()
+        control_path=os.path.join(args.generated_images_dir,names[0]+".png")
+        if (not os.path.exists(control_path)):
+            perceptual_model.set_reference_images(images_batch)
+            op = perceptual_model.optimize(generator.dlatent_variable, iterations=args.iterations, learning_rate=args.lr)
+            pbar = tqdm(op, leave=False, total=args.iterations)
+            for loss in pbar:
+                pbar.set_description(' '.join(names)+' Loss: %.2f' % loss)
+            print(' '.join(names), ' loss:', loss)
+    
+            # Generate images from found dlatents and save them
+            generated_images = generator.generate_images()
+            generated_dlatents = generator.get_dlatents()
+            for img_array, dlatent, img_name in zip(generated_images, generated_dlatents, names):
+                img = PIL.Image.fromarray(img_array, 'RGB')
+                img.save(os.path.join(args.generated_images_dir, f'{img_name}.png'), 'PNG')
+                np.save(os.path.join(args.dlatent_dir, f'{img_name}.npy'), dlatent)
+    
+            generator.reset_dlatents()
 
 
 if __name__ == "__main__":
